backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pyswisseph>=2.10.3
pytz==2023.3
fpdf==1.7.2
plotly==5.18.0
google-genai==0.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
cors
python-multipart
